# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Tour Trip API
  description: REST API for tour booking, payment, invoice, receipt, and participant management
  version: 1.0.0
servers:
  - url: http://localhost:8081/api/v1
paths:
  /bookings:
    post:
      tags: [Booking]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingApiResponse' }
    get:
      tags: [Booking]
      summary: List all bookings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingListApiResponse' }
  /bookings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Booking]
      summary: Get a booking by id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingApiResponse' }
        '404': { description: Not found }
    put:
      tags: [Booking]
      summary: Update a booking (userId, tourScheduleId, specialRequests, bookingType only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingApiResponse' }
        '404': { description: Not found }
    delete:
      tags: [Booking]
      summary: Soft-delete a booking
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VoidApiResponse' }
        '404': { description: Not found }
  /participants:
    post:
      tags: [Participant]
      summary: Add a participant to a booking
      parameters:
        - name: bookingId
          in: query
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ParticipantRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParticipantApiResponse' }
    get:
      tags: [Participant]
      summary: List all participants
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParticipantListApiResponse' }
  /participants/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Participant]
      summary: Get a participant by id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParticipantApiResponse' }
        '404': { description: Not found }
    put:
      tags: [Participant]
      summary: Update a participant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ParticipantRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParticipantApiResponse' }
        '404': { description: Not found }
    delete:
      tags: [Participant]
      summary: Soft-delete a participant
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VoidApiResponse' }
        '404': { description: Not found }
...
components:
  schemas:
    BookingType:
      type: string
      enum: [Individual, Family, Team]
    BookingStatus:
      type: string
      enum: [Pending, Confirmed, Cancelled]
    Sex:
      type: string
      enum: [Male, Female, Other]
...
    ParticipantRequest:
      type: object
      properties:
        name: { type: string }
        ageGroup: { type: string }
        sex: { $ref: '#/components/schemas/Sex' }
    ParticipantResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        ageGroup: { type: string }
        sex: { $ref: '#/components/schemas/Sex' }
    BookingRequest:
      type: object
      required: [userId, tourScheduleId, bookingType]
      properties:
        userId: { type: integer, format: int64 }
        tourScheduleId: { type: integer, format: int64 }
        specialRequests: { type: string }
        bookingType: { $ref: '#/components/schemas/BookingType' }
        participantRequests:
          type: array
          items: { $ref: '#/components/schemas/ParticipantRequest' }
    BookingResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        tourScheduleId: { type: integer, format: int64 }
        totalPrice: { type: number, format: decimal }
        specialRequests: { type: string }
        bookingStatus: { $ref: '#/components/schemas/BookingStatus' }
        bookingType: { $ref: '#/components/schemas/BookingType' }
        memberCount: { type: integer }
        participants:
          type: array
          items: { $ref: '#/components/schemas/ParticipantResponse' }
...
    VoidApiResponse:
      type: object
      properties:
        message: { type: string }
        status: { type: integer }
        data: { nullable: true }
    BookingApiResponse:
      type: object
      properties:
        message: { type: string }
        status: { type: integer }
        data: { $ref: '#/components/schemas/BookingResponse' }
    BookingListApiResponse:
      type: object
      properties:
        message: { type: string }
        status: { type: integer }
        data:
          type: array
          items: { $ref: '#/components/schemas/BookingResponse' }
    ParticipantApiResponse:
      type: object
      properties:
        message: { type: string }
        status: { type: integer }
        data: { $ref: '#/components/schemas/ParticipantResponse' }
    ParticipantListApiResponse:
      type: object
      properties:
        message: { type: string }
        status: { type: integer }
        data:
          type: array
          items: { $ref: '#/components/schemas/ParticipantResponse' }
...